Update release notes, add tests for match
