+++ tests/schema/test_arguments.py
@@ -125,53 +125,6 @@ def test_optional_argument_unset():
     assert result.data == {"hello": "Hi there"}
 
 
-def test_optional_argument_maybe() -> None:
-    @strawberry.type
-    class Query:
-        @strawberry.field
-        def hello(self, name: strawberry.Maybe[str] = None) -> str:
-            if name:
-                return "None" if name.value is None else name.value
-            return "UNSET"
-
-    schema = strawberry.Schema(query=Query)
-
-    assert str(schema) == dedent(
-        """\
-        type Query {
-          hello(name: String): String!
-        }"""
-    )
-
-    result = schema.execute_sync(
-        """
-        query {
-            hello
-        }
-    """
-    )
-    assert not result.errors
-    assert result.data == {"hello": "UNSET"}
-    result = schema.execute_sync(
-        """
-        query {
-            hello(name: "bar")
-        }
-    """
-    )
-    assert not result.errors
-    assert result.data == {"hello": "bar"}
-    result = schema.execute_sync(
-        """
-        query {
-            hello(name: null)
-        }
-    """
-    )
-    assert not result.errors
-    assert result.data == {"hello": "None"}
-
-
 def test_optional_input_field_unset():
     @strawberry.input
     class TestInput:
